Minor refactoring:
 I moved the code from AtomParselet into NumberParselet and deleted AtomParselet because it is not used by any other class.
